requirements.txt
flask>=2.2.5
requests==2.26.0
python-dotenv==0.19.0
flask-cors==3.0.10
amadeus==8.1.0
beautifulsoup4==4.9.3
selenium>=4.9.0
webdriver_manager==3.8.0
fake-useragent==1.1.1
gunicorn==20.1.0 